--- a/Macros/e3 Includes/e3_Assists.inc
+++ b/Macros/e3 Includes/e3_Assists.inc
@@ -261,6 +261,8 @@ SUB check_AssistStatus
 |- e3_casting is not used, for performance, as these are short recast low impact only
 |------------------------------------------------|
 SUB CombatAbilities
+
+  
   /if (${Debug} || ${Debug_Assists}) /echo |- CombatAbilities ==>
   /declare i int local
   | Prereq/priorities
@@ -282,10 +284,11 @@ SUB CombatAbilities
           |   /doability "Bazu Bellow"
           } else /if (${Me.AltAbilityReady[Divine Stun]}) {
             /bc Divine Stun on ${Target.CleanName}: ${Me.TargetOfTarget.Class.ShortName}-${Me.TargetOfTarget} has aggro
-            /casting "Divine Stun"
+            /call castSimpleSpell "Divine Stun" ${Target.ID}
+
           } else /if (${Me.SpellReady[Terror of Discord]}) {
             /bc Terror of Discord on ${Target.CleanName}: ${Me.TargetOfTarget.Class.ShortName}-${Me.TargetOfTarget} has aggro
-            /casting "Terror of Discord"
+            /call castSimpleSpell "Terror of Discord" ${Target.ID}
           }
         }
       }
@@ -302,18 +305,22 @@ SUB CombatAbilities
       /goto :skipAbility
     }
     /if (${Abilities2D[${i},${iCastType}].Equal[Ability]}) {
+
       /if (${Abilities2D[${i},${iCastName}].Equal[Bash]}) {
         /if (!${Select[${Me.Inventory[Offhand].Type},Shield]} && !${Me.AltAbility[2 Hand Bash]}) /goto :skipAbility
+        |can't do anything if stunned.
+        /if (${Me.Stunned}) /goto :skipAbility
+        /if (${Target.Distance} >10) /goto :skipAbility
+        
       }
       /if (${Me.AbilityReady[${Abilities2D[${i},${iCastName}]}]}) {
-        /doability "${Abilities2D[${i},${iCastName}]}"
-        /delay 2 !${Me.AbilityReady[${Abilities2D[${i},${iCastName}]}]}
+        
+        /call castSimpleSpell "${Abilities2D[${i},${iCastName}]}" 0
       }
       |- AA, only those with 0 cast time can go here
     } else /if (${Abilities2D[${i},${iCastType}].Equal[AA]}) {
       /if (${Me.AltAbilityReady[${Abilities2D[${i},${iCastName}]}]}) {
-        /casting "${Abilities2D[${i},${iCastName}]}" alt
-        /delay 3 !${Me.AltAbilityReady[${Abilities2D[${i},${iCastName}]}]}
+        /call castSimpleSpell "${Abilities2D[${i},${iCastName}]}" 0
       }
       |- Discipline, no disc with a duration can go here
     } else /if (${Abilities2D[${i},${iCastType}].Equal[Disc]}) {
@@ -322,8 +329,8 @@ SUB CombatAbilities
         |/echo ${Abilities2D[${i},${iCastName}]} myend ${Me.PctEndurance} >= minend ${Abilities2D[${i},${iMinEnd}]} and endcost ${Spell[${Abilities2D[${i},${iCastName}]}].EnduranceCost} <= ${Me.Endurance}
         /if (${Me.PctEndurance} >= ${Abilities2D[${i},${iMinEnd}]} && ${Spell[${Abilities2D[${i},${iCastName}]}].EnduranceCost} <= ${Me.Endurance} && ${Abilities2D[${i},${iIfs}]}) {
           |/echo ready ${abilityName} ${Me.CombatAbilityReady[${abilityName}]}
-          /disc ${Abilities2D[${i},${iCastName}]}
-          /delay 3 !${Me.CombatAbilityReady[${Abilities2D[${i},${iCastName}]}]}
+          /call castSimpleSpell "${Abilities2D[${i},${iCastName}]}" 0
+          
         }
       }
     }
@@ -427,14 +434,7 @@ SUB check_Nukes
       }
     } else {
 
-      |When an item delay is applied to an entry, check to see if the delay is still active
-      |Check lower comment after the spell is cast for more information.
-      /if (${Defined[nukeSpellRecast_${Nukes2D[${castIndex},${iSpellID}]}]}) {
-        /if (${nukeSpellRecast_${Nukes2D[${castIndex},${iSpellID}]}}) {
-          |/bc check_nuke: skipping nukes as its timer isn't met yet:${Nukes2D[${castIndex},${iCastName}]} time left:${nukeSpellRecast_${Nukes2D[${castIndex},${iSpellID}]}}
-          /goto :skipCast
-        }
-      }
+     
       | skip casting if a delay is defined and active
       /if (${Defined[castDelay${Nukes2D[${castIndex},${iCastID}]}]}) {
         /if (${castDelay${Nukes2D[${castIndex},${iCastID}]}}) {
@@ -467,13 +467,6 @@ SUB check_Nukes
         /if (!${Nukes2D[${i},${iCastName}].Equal[${currentCastName}]}) {
           
           /call check_Ready "Nukes2D" ${i}
-          |When an item delay is applied to an entry, check to see if the delay is still active
-          |Check lower comment after the spell is cast for more information.
-          /if (${Defined[nukeSpellRecast_${Nukes2D[${i},${iSpellID}]}]}) {
-            /if (${nukeSpellRecast_${Nukes2D[${i},${iSpellID}]}}) {
-              /varset c_Ready FALSE
-            }
-          }
           /if (${c_Ready} && ${Nukes2D[${i},${iIfs}]}) {
             /if (${Bool[${Nukes2D[${i},${iGiftOfMana}]}]}) {
               |its specified to use gift of mana, well do we have it?
@@ -580,16 +573,9 @@ SUB check_Nukes
       }
       
 
-      /if (${Select[${castReturn},CAST_SUCCESS]} && ${typeOfCast.Equal[Item]}) {
-          |when an item is click/cast, the server sends over the cooldown of said item
-          |so for a breif time between clicking there is an 'unkown' state where the items
-          |cooldown is still 0, even tho it isn't valid. Noticed this when using molten orbs.
-          |Adding a small delay to any item that is cast to prevent it from being a valid target
-          |on the next event cycle as event cycles in e3 are far faster than the server can send us
-          |the cooldown information
-          /call CreateTimer nukeSpellRecast_${Nukes2D[${castIndex},${iSpellID}]} "1s"
-      }
+     
       |if the nuke was succesful, see if there is a delay defined and create a timer which locks out all nukes for the duration
+      |mainly used for PBAOE stuns
       /if (${Bool[${Nukes2D[${castIndex},${iDelay}]}]} && ${Select[${castReturn},CAST_SUCCESS]}) {
         /if (${Debug} || ${Debug_Assists}) /echo delaying ${Nukes2D[${castIndex},${iCastName}]} for ${Nukes2D[${castIndex},${iDelay}]}
         /call CreateTimer "castDelay${Nukes2D[${castIndex},${iCastID}]}" "${Nukes2D[${castIndex},${iDelay}]}"
@@ -1135,8 +1121,8 @@ SUB cast_damageOverTimeSpells(int TargetID, string ArrayName)
           /if (${${resistCount}} >= ${${ArrayName}[${i},${iMaxTries}]}) {
             /varset ${resistCount} 0
             /call createTimer "${spellTimer}" "1m"
-        }
-      }
+          }
+        } u
       | If the spell did not take hold.
       } else /if (${Cast.Stored.Name.Equal[${${ArrayName}[${i},${iSpellName}]}]} && ${castReturn.Equal[CAST_TAKEHOLD]} && !${Cast.Stored.Name.Equal[${spammableSpell}]}) {
         /if (${Defined[${resistCount}]}) {