--- a/Macros/e3 Includes/e3_Assists.inc
+++ b/Macros/e3 Includes/e3_Assists.inc
@@ -1442,7 +1442,7 @@ SUB useBurns(ArrayName)
         /if (${Me.ActiveDisc.ID} && ${${ArrayName}[${i},${iCastName}].Equal[Savage Spirit]}) /goto :skipCurrent
         /if (${Me.ActiveDisc.ID} && ${${ArrayName}[${i},${iCastType}].Equal[Disc]} && ${Bool[${${ArrayName}[${i},${iDuration}]}]}) {
            | /echo Currently in disc and duration > 0. Skiping ${${ArrayName}[${i},${iCastName}]}. Duration: ${Bool[${${ArrayName}[${i},${iDuration}]}]} 
-           /echo Burns-Skip: ${i} ${${ArrayName}[${i},${iCastName}]}
+           |/echo Burns-Skip: ${i} ${${ArrayName}[${i},${iCastName}]}
            /delay 1
            /goto :skipCurrent