Fairly large patch for E3 this time.


***NOTE***
With this update, Rangers please don't put trueshot in the nuke area, place it in the melee abilities as that is designed for Discs/AA/Abilities.


Core feature:

Looting has been significantly sped up.

Fixed ***most*** item click issues (Note, most of this work was before the item click spell reset patch on the server)

This took a ton of work, required an update to the bard players and the way most things were called inside of E3

There is now also an internal list of items that is used by E3 to determine cooldowns so that E3 will not try and cast something even if the Client incorrectly says you can.

New Features:
*	/BeforeSpell, /AfterSpell, these new tags can be used on nukes/heals/etc to fire off after or before a spell/item/aa is cast. The after/before must be a spell/item/aa.
	Usage: Quick Burn=Volatile Mana Blaze/AfterSpell|Harvest of Druzzil
	Result: Volatile Mana Blaze will cast. E3 will stay within the e3_cast section and immediately cast Harvest of Druzzil

	Usage2: Quick Burn=Bifold Focus of the Evil Eye/AfterSpell|Ether Flame/Ifs|NoTwinCastAndBurning
	Result: IF the condition is met, click item Bifold Focus of the Evil Eye and then immediately cast Ether Flame.

*	Autofire is now a valid range type for rangers to use. This will make your ranger act like a caster but they will auto shoot and face the target. It has a max range of 200.
	IE: Assist Type (Melee/Ranged/Off)=Autofire
	
*	We now have a new E3 bard player. This takes the the place of the old "Twist" player. the old "Twist" player has been moved to MQ2Twist.
	New bard player is about as performant as I can make in e3 (required a bit of modification to e3), handles item clicks easily, and has no set delay for songs. So 0.5 sec songs, 
	instant cast items, etc are no problem for it. No gem configuring for items, just put the item name in the song location and it will work. 
	IE: 
	SongPlayer=Twist
	[main Melody]
	Song=Lute of the flowing Waters
	Song=Selo's Accelerating Chorus/Gem|1
	Song=Storm Blade/Gem|2
	Song=Psalm of Veeshan/Gem|3
	Song=Niv's Harmonic/Gem|4
	Song=Ancient: Call of Power/Gem|5

*	nowCast is a new feature, its like queueCast but far more aggressive.
	Example: 
	/bct nowCast <SpellName>
	/bc nowCast <SpellName> (/bc is for backward compatability with queueCast, use bct)
	/bct nowCast <SpellName> targetid=${Target.ID}
	/bct nowCast <SpellName> targetid=${Me.ID}
	
	This is a pretty useful feature, it will interrupt anything and do this NOW unless its a healer class casting a heal or has /NoInterrupt flag.
	This has also been updated for the new bard player, so that you can nowCast bard songs as well. IE: You wish to pulse a slow song real quick without interrupting your twist.

*	/NoInterrupt is now a thing.
	Usage: Pet Buff=Ancient: Burnout Blaze/NoInterrupt
	
	This will blocks all types of interrupts, nowCast, heal checks, etc.
	This change requires a discussion on how E3 and healing works.

	There are different type of interrupts.  However based on cases on how you are playing EQ and what stage you are at you need to deal with them differently. 

	An example is, My target just got healed and is now over 95%, should i interrupt my heal and conserve mana?

	Well, in early levels.. that is totally the case, it *is* a total waste of mana to heal someone who just got nearly healed to full as your heal didn't hit in time. 

	In later levels, where your tanks HP can yoyo up and down, it makes far more sense to overheal than to interrupt.  
	Reason? Wait 2 sec and suddenly your tank is now 30% hp because you stopped a heal that would have landed to heal him to 70% but you canceled it. 

	Now you can put /NoInterrupt next to the heals you wish to allow them to continue and just 'overheal' at worst case and 'clutch heal' at best 

	Another example is when Shamans need to get that Slow off but the other debuffs can wait if the tank needs a heal.
	put a /NoInterrupt next to their malo/slow, so that gets in ASAP then back to the tank, as a slow is just a fancy way of preventing damage which is like a fancy way of healing.

	I'm sure you can find many cases where you don't want to stop whatever your doing and either heal the tank, or save mana as the tank is fine. 

	The choice is up to you.

CORE CHANGE:

*	check_Heals is now #1 for every healer. This is a hidden #1 and was needed because of interrupts and the way that the "Check_" functions work. 
	They do not go 1,2,3,4,5. It can very much go 1,1,1,1,1,2,2,2,2,2,1,3 depending on actions taken, as an action will reset the loop every time.
	
Updates:
*	Burns updated to fire much faster now. Cleric/Druid/Shaman have kickouts so they continue healing however. (tank/xtargets). You may use /NoInterrupt on burns as well. 
	**NOTE** each burn takes about 0.3 seconds, if your healers have a lot of burns and you have /NoInterrupt, your tank may die.

*	Melody Pause/Resume now work.

*	When casting an item for MQ2Twist, it will now stop the song, click the item, and continue where it left off. While not ideal, it prevents the entire twist from restarting.
	**NOTE* this will need to be reviewed with items with less than 0.5 seconds as they can be cast while singing now.
	
Bug Fixes:
* 	Slam now works as long as you put it in the 1-6 ability button areas
*	Fixed Long standing bug in E3 concerning discs and if they were in use or not. (what helped burns out eventually)
*	Fixed a bug on HoT's being case not being interrupted if the tank/xtarget needed heals.

*	Partially fixed meming of spells. Were you would mem a spell... unmem, mem another spell.. repeat, this was needed for summoned items that were being cast out of combat.
	Note, if you are OOM this will still happen. Will hopefully address next patch.

Performance:

*	Performance enhancements for castSimpleSpell, which translates to most things cast in e3 since now everything uses it, that isn't one of the primary array types (dots/nukes/debuffs/heals/etc).

*	Routed most casting through E3_Cast which allowed us to programmatically prevent most spell gem lockups. 

*	Added a method to automatically detect and unlock spell gems when it does occur. 

	**Note**: This feature uses Origin / Marr's Calling with an interrupt. You will see your characters dismount if they're mounted. We've been using this for a couple weeks and no one has actually completed an Origin or Marr's Calling cast, but you're welcome to change the AA casted inside check_SpellGemLockup within the E3_Casting.inc file.**

*	Logging updates


**NOTE** Many of these items were addressed with Source Code updates that happened after we had already created a "fix" in E3. We did testing and found that while things like "spell gem" lockouts occur significant less often--they still occur. We left the E3 fix in place to automatically resolve the issue when it does occur.